C#: Move the generator class to a helper file.
